backend/requirements.txt
django>=5.1.0,<5.2.0
djangorestframework>=3.15.0,<3.16.0
django-cors-headers>=4.0.0,<4.8.0
dj-rest-auth>=7.0.0,<7.1.0
psycopg2-binary>=2.9.0,<2.10.0
pillow>=11.0.0,<11.2.0
python-dotenv>=1.0.0,<1.1.0
pyjwt>=2.6.0,<2.11.0
anthropic>=0.49.0,<0.50.0
boto3>=1.37.0,<1.38.0
# LangChain dependencies
langchain>=0.1.0,<0.2.0
langchain-anthropic>=0.1.0,<0.2.0
langchain-core>=0.1.0,<0.2.0
langsmith>=0.0.75,<0.1.0
pydantic>=2.6.0,<2.7.0
faiss-cpu>=1.7.4,<1.8.0